243: move GetLogsParam
