+++ src/rpc_client/requests.rs
@@ -1,6 +1,7 @@
-use crate::rpc_client::eth_rpc::BlockSpec;
+use crate::rpc_client::eth_rpc::{BlockSpec, FixedSizeData};
 use ic_ethereum_types::Address;
 use serde::Serialize;
+use candid::Deserialize;
 
 /// Parameters of the [`eth_getTransactionCount`](https://ethereum.org/en/developers/docs/apis/json-rpc/#eth_gettransactioncount) call.
 #[derive(Debug, Serialize, Clone)]
@@ -16,4 +17,22 @@ impl From<GetTransactionCountParams> for (Address, BlockSpec) {
     fn from(params: GetTransactionCountParams) -> Self {
         (params.address, params.block)
     }
+}
+
+/// Parameters of the [`eth_getLogs`](https://ethereum.org/en/developers/docs/apis/json-rpc/#eth_getlogs) call.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct GetLogsParam {
+    /// Integer block number, or "latest" for the last mined block or "pending", "earliest" for not yet mined transactions.
+    #[serde(rename = "fromBlock")]
+    pub from_block: BlockSpec,
+    /// Integer block number, or "latest" for the last mined block or "pending", "earliest" for not yet mined transactions.
+    #[serde(rename = "toBlock")]
+    pub to_block: BlockSpec,
+    /// Contract address or a list of addresses from which logs should originate.
+    pub address: Vec<Address>,
+    /// Array of 32 Bytes DATA topics.
+    /// Topics are order-dependent.
+    /// Each topic can also be an array of DATA with "or" options.
+    #[serde(skip_serializing_if = "Vec::is_empty")]
+    pub topics: Vec<Vec<FixedSizeData>>,
 }
